+++ src/nodes/base_node.rs
@@ -10,11 +10,13 @@ use crate::{
     Allocator,
     error::ArtError,
     lock::{ReadGuard, TypedReadGuard},
-    node_4::{Node4, Node4Iter},
-    node_16::{Node16, Node16Iter},
-    node_48::{Node48, Node48Iter},
-    node_256::{Node256, Node256Iter},
-    node_ptr::{AllocatedNode, NodePtr},
+    nodes::{
+        AllocatedNode, NodePtr,
+        node_4::{Node4, Node4Iter},
+        node_16::{Node16, Node16Iter},
+        node_48::{Node48, Node48Iter},
+        node_256::{Node256, Node256Iter},
+    },
 };
 
 pub(crate) const MAX_KEY_LEN: usize = 8;
