+++ src/nodes/node_256.rs
@@ -1,4 +1,4 @@
-use crate::{
+use super::{
     base_node::{BaseNode, Node, NodeIter, NodeType},
     node_ptr::NodePtr,
 };
