+++ src/nodes/node_ptr.rs
@@ -1,7 +1,7 @@
 use std::{fmt::Debug, ptr::NonNull};
 
 use crate::{
-    base_node::{BaseNode, Node},
+    nodes::{BaseNode, Node},
     utils::KeyTracker,
 };
